# net-ai-beginer/restaurant-booking-sample/RestaurantBookingChat/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Restaurant Booking API",
    "description": "Demo API for AI enabled app learning",
    "version": "v1"
  },
  "paths": {
    "/RestaurantBooking/AvailableTable": {
      "get": {
        "tags": [
          "RestaurantBooking"
        ],
        "parameters": [
          {
            "name": "reservationTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "capacity",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TableModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/RestaurantBooking/ReserveTable": {
      "post": {
        "tags": [
          "RestaurantBooking"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTableRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTableRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReserveTableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ReserveTableRequest": {
        "type": "object",
        "properties": {
          "tableId": {
            "type": "integer",
            "format": "int32"
          },
          "reservationTime": {
            "type": "string",
            "format": "date-time"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "customerPhone": {
            "type": "string",
            "nullable": true
          },
          "numberOfPeople": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "TableModel": {
        "required": [
          "description",
          "seats"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seats": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}